merge master to dev

+++ src/main/java/com/kiran/service/integration/YelpAPI.java
@@ -50,7 +50,7 @@ public class YelpAPI {
                 oneRestaurantInfo.put("name", jArray.getJSONObject(i).getString("name"));
                 oneRestaurantInfo.put("url", jArray.getJSONObject(i).getString("url"));
                 oneRestaurantInfo.put("review", Integer.toString(jArray.getJSONObject(i).getInt("review_count")));
-                oneRestaurantInfo.put("rating", Integer.toString(jArray.getJSONObject(i).getInt("rating")));
+                oneRestaurantInfo.put("rating", Double.toString(jArray.getJSONObject(i).getDouble("rating")));
                 oneRestaurantInfo.put("image_url", jArray.getJSONObject(i).getString("image_url"));
                 oneRestaurantInfo.put("location", jArray.getJSONObject(i).getJSONObject("location").getString("city"));
                 long distanceMeter = jArray.getJSONObject(i).getLong("distance");
